add authorization

+++ Controllers/ProductController.cs
@@ -1,5 +1,7 @@
-﻿using Ecommerce_System.Ecommerce.Application.InterfacesServices;
+﻿using Ecommerce_System.Ecommerce.Application.Helper;
+using Ecommerce_System.Ecommerce.Application.InterfacesServices;
 using Ecommerce_System.Ecommerce.Domain.DTO;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,10 +18,18 @@ namespace Ecommerce_System.Controllers
         }
 
         [HttpPost("AddProduct")]
+        [Authorize(Roles = "Admin")]
+
         public async Task<IActionResult> AddProduct([FromBody]ProductDto product)
         {
             try
             {
+                var userId = User.GetUserId();
+                if (userId == null)
+                {
+                    return Unauthorized();
+
+                }
                 var result = await _productService.CreateProductAsync(product);
                 return Ok(result);
             }
@@ -29,10 +39,18 @@ namespace Ecommerce_System.Controllers
             }
         }
         [HttpGet("GetAllProduct")]
+        [Authorize(Roles = "Admin , User")]
+
         public async Task<IActionResult> GetAllProduct()
         {
             try
             {
+                var userId = User.GetUserId();
+                if (userId == null)
+                {
+                    return Unauthorized();
+
+                }
                 var result = await _productService.GetAllProductAsync();
                 if (result == null)
                 {
@@ -49,10 +67,17 @@ namespace Ecommerce_System.Controllers
 
         }
         [HttpPost("UpdateProduct")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult>UpdateProduct(int id, [FromBody] ProductDto product)
         {
             try
             {
+                var userId = User.GetUserId();
+                if (userId == null)
+                {
+                    return Unauthorized();
+
+                }
                 var result = await _productService.UpdateProductAsync(id,product);
                 return Ok("product Updated successfly");
             }
@@ -62,10 +87,18 @@ namespace Ecommerce_System.Controllers
             }
         }
         [HttpDelete("DeleteProduct")]
+        [Authorize(Roles = "Admin")]
+
         public async Task<IActionResult>DeleteProduct(int id)
         {
             try
             {
+                var userId = User.GetUserId();
+                if (userId == null)
+                {
+                    return Unauthorized();
+
+                }
                 await _productService.DeleteProductAsync(id);
                 return Ok("Product Deleted Successfly");
             }
@@ -75,11 +108,18 @@ namespace Ecommerce_System.Controllers
             }
         }
         [HttpGet("GetProductById")]
+        [Authorize(Roles = "Admin , User")]
         public async Task<IActionResult>GetProductById(int id)
         {
             try
             {
-               var result= await _productService.GetById(id);
+                var userId = User.GetUserId();
+                if (userId == null)
+                {
+                    return Unauthorized();
+
+                }
+                var result= await _productService.GetById(id);
                 return Ok(result);
             }
             catch (Exception ex)
